Rename dir to RemoteMedia
